fix(core): skip min proposal duration fetch until space id is known
Proposal detail and related flows passed null/undefined spaceId into
useSpaceMinProposalDuration via unsafe casts, causing viem readContract
to throw and the root error boundary to show the generic maintenance UI.

+++ packages/epics/src/agreements/components/create-agreement-base-fields.tsx
@@ -163,7 +163,7 @@ export function CreateAgreementBaseFields({
   });
 
   const { duration } = useSpaceMinProposalDuration({
-    spaceId: spaceIdBigInt as bigint,
+    spaceId: spaceIdBigInt,
   });
 
   const { theme } = useTheme();
